requirements.txt
streamlit>=1.28.0
torch==2.7.0
langchain>=0.1.0
openai>=1.0.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
beautifulsoup4>=4.12.0
unstructured>=0.10.0
python-dotenv>=1.0.0
requests>=2.31.0
ragatouille>=0.0.7
rank_bm25
pandas>=1.5.0
openpyxl>=3.0.0